feat: display the number of different denominations of stamps in the aside section
* feat: display the number of different denominations of stamps in the aside section

* feat: add calculation of number of combinations based on number of stamps and exponent value

The changes in this commit add a calculation for the number of combinations based on the number of different stamp denominations and the exponent value. This calculation is displayed in the UI and is updated dynamically when the number of stamps or the exponent value changes. This provides users with more information about the potential number of combinations they can create with their stamps.