chore: update styling of the main section and add box shadow and border radius to the aside section (#18)
* chore: update styling of the main section and add box shadow and border radius to the aside section

* Merge branch 'master' into add-legend-to-stamp-list

* Merge pull request #1 from ilenia-magoni/add-legend-to-stamp-list

Add legend to stamp list

* Improve styling and layout (#2)

* style: update styling for main, fieldset, and body elements in login page (#3)

* feat: display the number of different denominations of stamps in the aside section (#4)

* feat: display the number of different denominations of stamps in the aside section

* feat: display the number of different denominations of stamps in the aside section

* feat: add calculation of number of combinations based on number of stamps and exponent value

The changes in this commit add a calculation for the number of combinations based on the number of different stamp denominations and the exponent value. This calculation is displayed in the UI and is updated dynamically when the number of stamps or the exponent value changes. This provides users with more information about the potential number of combinations they can create with their stamps.

* fix: fix calculation of number of combinations displayed in UI to show 2 decimal places (#6)

* fix: validate stamp quantity as a whole number before calculating value (#7)

* feat: improve calculation of number of combinations (#8)

- Refactored the code to use a separate function `numberCombinations` to calculate the number of combinations.
- Added a condition to format the number in exponential notation if it exceeds 1000.

* Update number combinations function (#9)

* feat: improve calculation of number of combinations

- Refactored the code to use a separate function `numberCombinations` to calculate the number of combinations.
- Added a condition to format the number in exponential notation if it exceeds 1000.

* fix: update condition for numberCombinations function to return exponential notation for numbers greater than 9999

* feat: display the combination of stamps to be used for postage (#10)

The code now displays the combination of stamps to be used for postage in the `usaQuestiFrancobolli` element. The combination is generated by mapping the `face_value` property of each stamp in the `currentSet` array and joining them with a comma.

* fix: filter stamps based on postage + range before generating combinations (#11)

The changes were made to fix an issue in the code where the stamps combinations were not being generated correctly. The code now correctly filters the stamps based on the postage + range before generating the combinations, ensuring that only stamps with values less than or equal to the postage + range are considered in the combinations. This fix ensures that the correct combinations of stamps are generated and used in the calculations.

* refactor: remove console.log statement (#12)

* chore: refactor button container in stamp calculator UI (#13)

* Fix conflict in retrieve stamps function (#14)

* style: add total value column to stamp list table (#15)

* fix: correct spanTotal assignment and display in UI (#16)

* fix: display message when no stamps are available for combination (#17)

* fix: set innerHTML for spanTotal correctly in case of no possible combinations