feat: Add radio button to select stamp to use (#21)
Added radio buttons to each stamp in the stamp list table to allow the user to select a specific stamp to use for postage calculation. When a radio button is selected, the selected stamp is stored in the variable francobollo_da_usare for further processing. This enhancement provides more control and customization options for the user when calculating postage.